- add missing files and re-org DllFinder testing classes into TestUtil
